move h5py binary helpers to io_utils, integrate in load/save_model
